Realocate class info inside class activity
