--- app/src/main/res/layout/fragment_show_class_components.xml
@@ -1,16 +1,0 @@
-<FrameLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="fga.mds.gpp.trezentos.View.Fragment.ShowClassComponentsFragment"
-    android:background="#ffbd19">
-
-    <android.support.v7.widget.RecyclerView
-        android:id="@+id/recycler_class_components"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:clipToPadding="false"
-        />
-
-</FrameLayout>
